实现回复功能-异常处理

+++ src/main/java/life/fofo/community/controller/PublishController.java
@@ -1,7 +1,6 @@
 package life.fofo.community.controller;
 
 import life.fofo.community.dto.QuestionDTO;
-import life.fofo.community.mapper.QuestionMapper;
 import life.fofo.community.model.Question;
 import life.fofo.community.model.User;
 import life.fofo.community.service.QuestionService;
@@ -22,7 +21,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping(value = "/publish/{id}")
-    public String edit(@PathVariable(name = "id")Integer id,Model model){
+    public String edit(@PathVariable(name = "id")Long id,Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
         model.addAttribute("description", question.getDescription());
@@ -41,7 +40,7 @@ public class PublishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "id", required = false) Integer id,
+            @RequestParam(value = "id", required = false) Long id,
             HttpServletRequest request,
             Model model) {
         model.addAttribute("title", title);
